refactor: extract orphanAttributable into shared orphan-attribution.ts
The migrate orphan-sweep (sweepOrphanOnPort) and the supervisor's
crash-restart adopt-kill must answer the SAME safety-critical question
before signalling a port-holder: is this process plausibly THIS parachute
module (adoptable) or an operator's unrelated process (never touch)?

Move orphanAttributable + the OwnerProbeFn type out of
commands/migrate-cutover.ts into a new src/orphan-attribution.ts so both
sites share ONE implementation — a loosened needle can't widen the kill
surface in one place without the other noticing. migrate-cutover.ts
re-exports OwnerProbeFn for its existing import surface; behavior is
byte-identical (the migrate tests are unchanged).
